Factor InstNamer out of the formatter.
